# openapi.yaml
openapi: 3.0.0
servers:
  - url: https://api.kalastud.io
    description: Production Server
  - url: https://perfect-week-test.herokuapp.com
    description: Test Server
info:
  title: PerfectWeek API
  description: The PerfectWeek web API
  version: 0.1.0
  contact:
    email: perfectweek_2020@labeip.epitech.eu
  license:
    name: MIT
    url: 'https://github.com/PerfectWeek/web-api/blob/dev/LICENSE'

####################
# Tags
#
...
paths:
...
  /calendars/{calendar-id}/assistant/find-best-slots:
    get:
      tags:
        - Calendar
      summary: find best time slots for given Event
      operationId: findBestSlots
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: calendar-id
          required: true
          description: The Calendar id
          schema:
            type: number
        - in: query
          name: duration
          required: true
          description: The event duration, in minutes
          schema:
            type: number
        - in: query
          name: location
          required: false
          description: The event location
          schema:
            type: string
        - in: query
          name: min_time
          required: true
          description: The event minimum start time (yyyy-mm-ddThh:mm:ss)
          schema:
            type: string
            format: date-time
        - in: query
          name: max_time
          required: true
          description: The event maximum end time (yyyy-mm-ddThh:mm:ss)
          schema:
            type: string
            format: date-time
        - in: query
          name: type
          required: true
          description: The type of event to be created (party, work, hobby, workout)
          schema:
            type: string
        - in: query
          name: limit
          required: false
          description: The max number of timeslots to be returned
          schema:
            type: number

      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: OK
                  slots:
                    type: array
                    items:
                      $ref: '#/components/schemas/SlotInfo'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationFailed'
        '403':
          description: Action not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
  /calendars/{calendar-id}/assistant/get-event-suggestions:
    get:
      tags:
        - Calendar
      summary: Request event suggestions
      operationId: getEventSuggestions
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: calendar-id
          required: true
          description: The Calendar id
          schema:
            type: number
        - in: query
          name: min_time
          required: true
          description: The event minimum start time (yyyy-mm-ddThh:mm:ss)
          schema:
            type: string
            format: date-time
        - in: query
          name: max_time
          required: true
          description: The event maximum end time (yyyy-mm-ddThh:mm:ss)
          schema:
            type: string
            format: date-time
        - in: query
          name: limit
          required: false
          description: The max number of events to be returned
          schema:
            type: number

      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: OK
                  slots:
                    type: array
                    items:
                      $ref: '#/components/schemas/EventSuggestion'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationFailed'
        '403':
          description: Action not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'


  #
  # Event
  #
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  #
  # Schemas
  #
  schemas:
...
    SlotInfo:
      type: object
      properties:
        start_time:
          type: string
          example: "2019-03-08T12:12:12"
        end_time:
          type: string
          example: "2019-03-08T15:12:12"
        score:
          type: number
          example: 0.9
    EventSuggestion:
      type: object
      properties:
        event:
          $ref: '#/components/schemas/EventInfo'
        score:
          type: number
          example: 0.9
    EventInfo:
      type: object
      properties:
        id:
          type: number
          example: 43
        name:
          type: string
          example: nouvel an
        description:
          type: string
          example: soiree
        location:
          type: string
          example: Le pub Universitaire, Quebec City
        type:
          type: string
          example: party
        visibility:
          type: string
          example: public
        calendar_id:
          type: number
          example : 3
        start_time:
          type: string
          example: "2012-12-12T12:12:12"
        end_time:
          type: string
          example: "2012-12-12T12:12:12"
        status:
          type: string
          example: invited
...
    AuthenticationFailed:
      description: The authentication failed
      type: object
      properties:
        message:
          type: string
          example: You must be authenticated to perform this action
    # 403
    Unauthorized:
      description: The action is not allowed
      type: object
      properties:
        message:
          type: string
          example: Action not allowed
    # 404
...
    InternalServerError:
      description: Internal Server Error
      type: object
      properties:
        message:
          type: string
          example: Internal Server Error